Version 0.9 	modifié:         install_en.sh 	modifié:         install_fr.sh

+++ install_en.sh
@@ -9,8 +9,10 @@ echo "cd src"
 cd src
 
 echo "Deleting previous compilation files"
+if [ -d .../build ]
+then
 rm ../build/*
-echo "qmake"
+fi
 
 qmake -o ../build/Makefile
 if [ $? -ne 0 ]
